// packages/react/src/models/branding-preferences.ts
export interface BrandingPreferenceInterface {

  /**
     * Configurations.
     */
  configs: BrandingPreferenceConfigInterface;
  /**
     * images such as Logo, Favicon, etc.
     * @deprecated Use the images object in `theme.[<DESIRED_THEME>].images`.
     */
  images?: BrandingPreferenceImagesInterface;
  /**
     * Layout.
     */
  layout: BrandingPreferenceLayoutInterface;
  /**
     * Organization's basic details.
     */
  organizationDetails: BrandingPreferenceOrganizationDetailsInterface;
  /**
     * Stylesheets for login pages etc..
     */
  stylesheets?: BrandingPreferenceStylesheetsInterface;
  /**
     * Theme.
     */
  theme: BrandingPreferenceThemeInterface;
  /**
     * Links for policies, etc.
     */
  urls: BrandingPreferenceURLInterface;
}

/**
 * Interface Branding preference organization details.
 */
export interface BrandingPreferenceOrganizationDetailsInterface {
  /**
     * Copyright for the footer.
     * @deprecated Moved to the `/branding-preference/text` API.
     */
  copyrightText?: string;
  /**
     * Display name to be shown for Org members.
     */
  displayName: string;
  /**
     * Site title appearing on the browser tab.
     * @deprecated Moved to the `/branding-preference/text` API.
     */
  siteTitle?: string;
  /**
     * Support email to be shown for Org members.
     */
  supportEmail: string;
}

/**
 * Interface Branding preference images.
 */
export interface BrandingPreferenceImagesInterface {
  /**
     * Organization Favicon.
     */
  favicon: Omit<BrandingPreferenceImageInterface, 'altText'>;
  /**
     * Organization Logo.
     */
  logo: BrandingPreferenceImageInterface;
  /**
     * Organization My Account Logo.
     */
  myAccountLogo: BrandingPreferenceImageInterface;
}

/**
 * Interface Branding preference image.
 */
export interface BrandingPreferenceImageInterface {
  /**
     * Image Alt.
     */
  altText: string | undefined;
  /**
     * Image URL.
     */
  imgURL: string | undefined;
  /**
     * Title.
     */
  title?: string;
}

/**
 * Interface Branding preference URLs.
 */
export interface BrandingPreferenceURLInterface {
  /**
     * Link for Cookie Policy.
     */
  cookiePolicyURL: string;
  /**
     * Link for Privacy Policy.
     */
  privacyPolicyURL: string;
  /**
     * Link for Terms of Service.
     */
  termsOfUseURL: string;
}

/**
 * Interface Branding preference stylesheets.
 */
export interface BrandingPreferenceStylesheetsInterface {
  /**
     * Login portal stylesheet.
     */
  accountApp: PredefinedThemes;
}

export type BrandingPreferenceThemeInterface = StrictBrandingPreferenceThemeInterface
& DynamicBrandingPreferenceThemeInterface;

/**
 * Interface Branding preference theme.
 */
export type DynamicBrandingPreferenceThemeInterface = {
  [ key in PredefinedThemesKeys ]: ThemeConfigInterface;
};
// ...
export interface StrictBrandingPreferenceThemeInterface {
  /**
     * The active theme.
     */
  activeTheme: PredefinedThemes;
}
// ...
export type BrandingPreferenceLayoutInterface = StrictBrandingPreferenceLayoutInterface
& Partial<DynamicBrandingPreferenceLayoutInterface>;

/**
 * Strict Interface Branding preference layout.
 */
export interface StrictBrandingPreferenceLayoutInterface {
  /**
     * The active layout.
     */
  activeLayout: PredefinedLayouts;
}

/**
 * Interface dynamic branding preference layout.
 */
export type DynamicBrandingPreferenceLayoutInterface =
    BrandingPreferenceSideImageLayoutInterface
    & BrandingPreferenceSideAlignedLayoutInterface;
// ...
export interface BrandingPreferenceConfigInterface {
  /**
     * Should the changes be published?
     */
  isBrandingEnabled: boolean;
  /**
     * Should remove default branding.
     * @deprecated Renamed to `removeDefaultBranding` to keep it common.
     */
  removeAsgardeoBranding?: boolean;
  /**
     * Should remove default branding.
     */
  removeDefaultBranding?: boolean;
}
// ...
export enum PredefinedThemes {
  DARK = 'DARK',
  LIGHT = 'LIGHT',
}